add more plugins for better editings

+++ setup.sh
@@ -11,30 +11,32 @@ mkdir -p "$HOME/.config"
 rm -rf "$HOME/.config/nvim"
 ln -s "$DOTFILES/nvim" "$HOME/.config/nvim"
 
-# --- 3. Tmux ---
-echo "Linking Tmux..."
-rm -rf "$HOME/.config/tmux"
-ln -s "$DOTFILES/tmux" "$HOME/.config/tmux"
-tmux source-file ~/.config/tmux/tmux.conf
+if [[ ! "$1" == "nix" ]]; then
+	# --- 3. Tmux ---
+	echo "Linking Tmux..."
+	rm -rf "$HOME/.config/tmux"
+	ln -s "$DOTFILES/tmux" "$HOME/.config/tmux"
+	tmux source-file ~/.config/tmux/tmux.conf
 
-# --- 3. AeroSpace ---
-echo "Linking AeroSpace..."
-rm -f "$HOME/.aerospace.toml"
-ln -s "$DOTFILES/aerospace/aerospace.toml" "$HOME/.aerospace.toml"
-aerospace reload-config
+	# --- 3. AeroSpace ---
+	echo "Linking AeroSpace..."
+	rm -f "$HOME/.aerospace.toml"
+	ln -s "$DOTFILES/aerospace/aerospace.toml" "$HOME/.aerospace.toml"
+	aerospace reload-config
 
-# --- 4. Zsh (Shell) ---
-echo "Linking Zshrc..."
-rm -f "$HOME/.zshrc"
-ln -s "$DOTFILES/zsh/.zshrc" "$HOME/.zshrc"
+	# --- 4. Zsh (Shell) ---
+	echo "Linking Zshrc..."
+	rm -f "$HOME/.zshrc"
+	ln -s "$DOTFILES/zsh/.zshrc" "$HOME/.zshrc"
 
-# --- 5. Scripts ---
-echo "Linking Scripts..."
-rm -rf "$HOME/scripts"
-ln -s "$DOTFILES/scripts" "$HOME/scripts"
+	# --- 5. Scripts ---
+	echo "Linking Scripts..."
+	rm -rf "$HOME/scripts"
+	ln -s "$DOTFILES/scripts" "$HOME/scripts"
 
-# SAFETY NET: Force all scripts to be executable
-echo "⚙️  Making scripts executable..."
-chmod +x "$DOTFILES/scripts/"*
+	# SAFETY NET: Force all scripts to be executable
+	echo "⚙️  Making scripts executable..."
+	chmod +x "$DOTFILES/scripts/"*
 
-echo "✅ All done! Please restart your terminal."
+	echo "✅ All done! Please restart your terminal."
+fi
